refactor(DekItem): move the package into ui.component
