+++ app/src/main/java/tcg/pocket/dex/tierdecks/TierDecksScreen.kt
@@ -8,6 +8,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import tcg.pocket.dex.component.DeckItem
 import tcg.pocket.dex.ui.theme.TcgPocketDexTheme
 
 @Composable
